Rename phone-number-placeholders.js to phone-numbers.js
